refactor(comment): CursorConvertible 인터페이스를 도메인 패키지로 이동
